# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciamento financeiro pessoal",
        "title": "API Financeiro",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "Victor Ramos",
            "email": "victor@email.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8085",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/ganhos": {
            "post": {
                "description": "Cria um novo registro de ganho com os dados fornecidos no corpo da requisição",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ganhos"
                ],
                "summary": "Cria um novo ganho",
                "parameters": [
                    {
                        "description": "Dados do ganho a ser criado",
                        "name": "ganho",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contracts.RequestGanhos"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ganho criado com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Erro nos dados enviados",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno no servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/gastos": {
            "post": {
                "description": "Cria um novo registro de gastos com os dados fornecidos no corpo da requisição",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gastos"
                ],
                "summary": "Cria um novo gasto",
                "parameters": [
                    {
                        "description": "Dados do gasto a ser criado",
                        "name": "gasto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contracts.RequestGastos"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Gasto criado com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Erro nos dados enviados",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno no servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ganhos/": {
            "get": {
                "description": "Fetches a ganho record based List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ganhos"
                ],
                "summary": "Retrieve a ganho list",
                "responses": {
                    "200": {
                        "description": "Ganho retrieved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Ganho not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ganhos/{id}": {
            "get": {
                "description": "Fetches a ganho record based on the provided ID parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ganhos"
                ],
                "summary": "Retrieve a ganho by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID of the ganho to retrieve",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ganho retrieved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Ganho not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/gastos/": {
            "get": {
                "description": "Fetches a gasto record based List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gastos"
                ],
                "summary": "Retrieve a gasto list",
                "responses": {
                    "200": {
                        "description": "Gasto retrieved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Gasto not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/gastos/{id}": {
            "get": {
                "description": "Fetches a gasto record based on the provided ID parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gastos"
                ],
                "summary": "Retrieve a gasto by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID of the gasto to retrieve",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gasto retrieved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Gasto not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "contracts.RequestGanhos": {
            "type": "object",
            "properties": {
                "categoria": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "observacao": {
                    "type": "string"
                },
                "tipo": {
                    "type": "string"
                },
                "usuario_id": {
                    "type": "integer"
                },
                "valor": {
                    "type": "number"
                }
            }
        },
        "contracts.RequestGastos": {
            "type": "object",
            "properties": {
                "categoria": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "data_inicio": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "forma_pagamento": {
                    "type": "string"
                },
                "observacao": {
                    "type": "string"
                },
                "parcelas": {
                    "type": "integer"
                },
                "usuario_id_created_at": {
                    "type": "integer"
                },
                "usuario_id_updated_at": {
                    "type": "integer"
                },
                "valor": {
                    "type": "number"
                }
            }
        }
    }
}